+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -1,4 +1,5 @@
 export * from './AnthropicParameters';
+export * from './AnthropicCompatibleParameters';
 export * from './AzureParameters';
 export * from './BedrockParameters';
 export * from './CerebrasParameters';